W-Line updates for trains that don't go to end
When looking at W-Line westbound trains, add legit to query the DB to
determine if the train will continue all the way to the last two stops
of the line (Red Rocks/JeffCo stations). If the train does not
continue, display an icon with a dot to notify the user.